Change variable ActuateCurrent from a float to an int

+++ march_hardware/include/march_hardware/Joint.h
@@ -47,7 +47,7 @@ public:
   void resetIMotionCube();
 
   void actuateRad(float targetPositionRad);
-  void actuateCurrent(float targetCurrentRad);
+  void actuateCurrent(int targetCurrentRad);
   int getActuationMode();
 
   float getAngleRad();
